requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
plotly
pandas
pdf2image
pillow
